로그인 로그아웃시 회원에 등록된 superRelate 데이터가 삭제되는 현상 때문에 조건식 추

+++ components/DynamicFields/SuperRelate/SuperRelateField.php
@@ -171,7 +171,9 @@ class SuperRelateField extends AbstractType
         $insertParam['s_id'] = $where['target_id'];
         $insertParam['s_group'] = $where['group'];
 
-        $this->handler->connection()->table(self::TABLE_NAME)->where($insertParam)->delete();    // all delete
+        if(array_get($args,'login_at') == null && array_get($args,'remember_token') == null) {
+            $this->handler->connection()->table(self::TABLE_NAME)->where($insertParam)->delete();    // all delete
+        }
 
         $insertParam['s_type'] = (strpos($config->get('group'), 'documents_') !== false) ? 'document' : 'user';
         $insertParam['t_group'] = $is_user ? 'user' : sprintf('documents_%s', $config->get('r_instance_id'));
